trying a new datatype

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,7 +10,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int k;
+	size_t k;
 
 	if (size == 0)
 		return;
